requirements.txt
aniso8601==9.0.1
blinker==1.6.2
click==8.1.3
colorama==0.4.6
Flask==2.3.2
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
marshmallow==3.19.0
packaging==23.1
psycopg2-binary==2.9.6
PyJWT==2.7.0
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
SQLAlchemy==2.0.13
typing_extensions==4.5.0
Werkzeug==2.3.4
Flask-JWT
